requirements.txt
azure-mgmt-resource==0.31.0
azure-mgmt-network==0.30.0